+++ R/int_gendeterm.R
@@ -7,9 +7,14 @@
 # @param dtSim Incomplete simulated data.table
 # @return A data.frame column  with the updated simulated data
 
-gendeterm <- function(n,formula,dtSim) {
+gendeterm <- function(n, formula, link, dtSim) {
+
   new <- with(dtSim,eval(parse(text = as.character(formula))))
   if (is.null(dtSim)) new = rep(new,n)
+
+  if (link=="log") new = exp(new)
+  else if (link == "logit") new = 1 / (1 + exp(-new))
+
   return(as.numeric(new))
 }