+++ NEWS.md
@@ -10,3 +10,9 @@
 * Increased speed required to generate categorical data with large sample sizes
 * Categorical data can now accomodate probabilities condition on covariates
 * Fix: package data.table 1.10.0 broke genMissDataMat. genMissDataMat has been updated.
+
+# simstudy 0.1.3
+
+* Modified "nonrandom" data generation to allow "log"" and "logit"" link options.
+* Added function genCorGen - generate a new data.table with correlated data from various distributions.
+* Added function addCorData - add correlated data from various distributions to existing data.tables.